refactor to separate evaluators from factory
