+++ application/backend/app/services/evaluation/__init__.py
@@ -1,0 +1,22 @@
+# Copyright (C) 2025 Intel Corporation
+# SPDX-License-Identifier: Apache-2.0
+
+from .evaluators import (
+    AveragingMethod,
+    DetectionEvaluator,
+    Evaluator,
+    InstanceSegmentationEvaluator,
+    MultiClassClassificationEvaluator,
+    MultiLabelClassificationEvaluator,
+)
+from .factory import EvaluatorFactory
+
+__all__ = [
+    "Evaluator",
+    "EvaluatorFactory",
+    "MultiClassClassificationEvaluator",
+    "MultiLabelClassificationEvaluator",
+    "DetectionEvaluator",
+    "InstanceSegmentationEvaluator",
+    "AveragingMethod",
+]
